+++ test/routing_test.rb
@@ -165,6 +165,21 @@ class RoutingTest < Minitest::Test
     assert_equal 404, status
   end
 
+  it "uses 404 error handler for not matching route" do
+    mock_app {
+      not_found do
+        "nf"
+      end
+      error 404 do
+        "e"
+      end
+    }
+
+    get "/"
+    assert_equal "e", body
+    assert_equal 404, status
+  end
+
   it 'matches empty PATH_INFO to "/" if no route is defined for ""' do
     mock_app do
       get '/' do
@@ -662,6 +677,23 @@ class RoutingTest < Minitest::Test
     get '/bar'
     assert ok?
     assert_equal 'Hello World', body
+  end
+
+  it "makes original request params available in error handler" do
+    mock_app {
+      disable :raise_errors
+
+      get '/:foo' do
+        raise ArgumentError, "foo"
+      end
+
+      error do
+        "Hello #{params['foo']}2"
+      end
+    }
+
+    get '/bar'
+    assert_equal 'Hello bar2', body
   end
 
   it "transitions to 404 when passed and no subsequent route matches" do